+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make `bakpack --version` print `bakpack X.Y.Z`, normalizing release tags like `vX.Y.Z` for display.
+
 ## [0.1.0] - 2026-06-23
 
 Initial release of `bakpack`.